test: move entires testing
